Window popups: don't use doubles when floats will do,
don't use floats when ints will do.

AbandonGameState: align correctly w/ PauseState window.
Ruleset: reformat.

UfopaediaSelectState
UfopaediaStartState
InterceptState
FundingState
	- center windows on the Globe.

GeoscapeState: cast windowPops tracker down to int.
- just 'cause

+++ src/Geoscape/GeoscapeState.cpp
@@ -2145,7 +2145,7 @@ void GeoscapeState::time10Minutes()
 
 	if (_windowPops > 0)
 	{
-		_ufoDetected->setText(Text::intWide(_windowPops));
+		_ufoDetected->setText(Text::intWide(static_cast<int>(_windowPops)));
 		_ufoDetected->setVisible();
 	}
 }
@@ -4528,7 +4528,7 @@ void GeoscapeState::assessUfoPopups()
 	if (--_windowPops == 0)
 		_ufoDetected->setVisible(false);
 	else
-		_ufoDetected->setText(Text::intWide(_windowPops));
+		_ufoDetected->setText(Text::intWide(static_cast<int>(_windowPops)));
 }
 
 /**
